Update lizzer_auto_shaders.py
Updated the getProperty function to have a backup option when handling the "Limbal Color" property since .to_dict() gave me an error but .to_list() seemed to work, so it'll now do that if .to_dict() throws an AttributeError. If that also fails, it raises a normal error.
Also the property data is now stored in a variable at the top of the function to reduce code repetition and re-reading the data.